@@ -4,31 +4,32 @@ description: GTU BT mobile app.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
-  flutter_bloc: 6.1.0
-  equatable: ^1.2.6
-  json_annotation: ^3.1.1
-  http: ^0.12.2
+  flutter_bloc: ^7.0.0
+  equatable: ^2.0.0
+  json_annotation: ^4.0.1
+  http: ^0.13.2
   flutter:
     sdk: flutter
-  firebase_core: ^0.7.0
-  firebase_auth: ^0.20.1
-  firebase_crashlytics: ^0.4.0+1
-  basic_utils: ^2.7.1
-  dartz: ^0.9.2
-  timeago: ^2.0.30
-  intl: ^0.16.1
-  another_flushbar: ^1.10.18
-  kiwi: ^2.1.1
-  cupertino_icons: ^1.0.0
-  image_picker: ^0.6.7+13
+  firebase_core: ^1.1.0
+  firebase_auth: ^1.1.3
+  firebase_crashlytics: ^2.0.2
+  basic_utils: ^3.0.2
+  dartz: ^0.10.0-nullsafety.1
+  timeago: ^3.0.2
+  intl: ^0.17.0
+  another_flushbar: ^1.10.20
+  kiwi: ^3.0.0-nullsafety.1
+  cupertino_icons: ^1.0.2
+  image_picker: ^0.7.5
+  collection: ^1.15.0-nullsafety.4
 
 dev_dependencies:
-  build_runner: ^1.11.1
-  json_serializable: ^3.5.1
-  flutter_native_splash: ^0.3.2
+  build_runner: ^2.0.2
+  json_serializable: ^4.1.1
+  flutter_native_splash: ^1.1.8+4
   flutter_test:
     sdk: flutter
 
