@@ -5,7 +5,7 @@ import 'package:image_picker/image_picker.dart';
 abstract class UserEvent extends Equatable {
   final String field;
 
-  UserEvent({@required this.field}) : super();
+  UserEvent({required this.field}) : super();
 
   @override
   List<Object> get props => [field];
@@ -13,7 +13,7 @@ abstract class UserEvent extends Equatable {
 
 class PhotoChanged extends UserEvent {
   final PickedFile imageFile;
-  PhotoChanged({@required this.imageFile}) : super(field: 'Photo');
+  PhotoChanged({required this.imageFile}) : super(field: 'Photo');
 
   @override
   String toString() => 'Profile Photo Changed';
