@@ -7,8 +7,8 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 class LoginPage extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
-
     return Scaffold(
+      resizeToAvoidBottomInset: false,
       backgroundColor: ColorSets.appMainColor,
       body: BlocProvider(
         create: (context) {
